backend/requirements.txt
# Backend requirements
fastapi>=0.95.0
uvicorn>=0.22.0
python-multipart>=0.0.6
git+https://github.com/illuin-tech/colpali
torch>=2.0.0
Pillow>=9.5.0
numpy>=1.24.0
pdf2image>=1.16.0
setuptools>=65.0.0
ollama>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0